Remove et_splices
Summary:
Working toward supporting nested expression trees, make the treatment of splices uniform between codegen and typechecking.

We remove the `et_splices` field from the expression tree AST node, and instead always insert the assignments of splices to temporaries at the first of the lambda generated during e.t. desugaring, essentially making the typechecker see the same code as codegen. There are still some things in typechecking that are sensitive to whether we are in an e.t. or not (e.g., is the `$this` variable valid). To allow us to track when we enter a splice (and hence leave an e.t.), we always keep the `ETSplice` node around the splice, but treat is semantically as the identity function. An extra field on it lets type type checker know if it should extract the client type (as in the virtualized expression) or not (as in the list of splice assignments to temporaries).

Some care is taken in IdentifySymbolService to just look at the splice assignment and type expression without providing the visitor expressions, which would just overwhelm hover with irrelevant stuff.

By moving the splices into the lambda inside of the runtime expression, we lose a little bit of completens.
  - Since the type environment is thrown away at the end of the lambda, any `as` refinements that happened inside of splices will now be lost after the expression tree the splices were in
  - Variables used in the splices are now treated as captured by the lambda, and might be pessimised

Reviewed By: dlreeves

Differential Revision: D54951681

fbshipit-source-id: 8bfd23f84f3dd100a741bf7b67b39f52091e3c14