Do not crash on missing fixup while producing PHP stack while crashing
Summary:
A common crash while crashing is when trying to obtain a PHP stacktrace where
the original crash occurred. If we crashed in jitted code, or a C++ helper that
does not have a fixup, VMRegAnchor inside createBacktrace() will not find a
fixup and crash.

Let's instead sync registers proactively in a soft mode and do not even attempt
to call createBacktrace() if there was no fixup. Instead, fall back to the
approximate logic instead.

This does not really change the behavior, as a crash in createBacktrace() would
cause reentrance of bt_handler(), which would fallback to the approximate logic
as well, but it makes debugging somewhat more pleasant, not having to skip a
bunch of frames where we crashed again. There's also less chance that something
else got corrupted due to the second crash.

Reviewed By: ricklavoie, mdko

Differential Revision: D56315850

fbshipit-source-id: 7980f9117796b27d4621876268e7ad93d6634b74