Start from RBP of the crash when determining fallback PHP stacktrace
Summary:
The frame pointer where the crash happened is available on X64. Use it when
looking up VM frames and skip frames associated with the crash handler.

I saw an example where the signal handler started inside the VM stack space
(which is obviously wrong) and we incorrectly determined that these frames are
VM frames and crashed when trying to parse them as ActRecs. Unable to
investigate this further due to lldb's TLS issues, but this diff should at
least avoid the issue for now.

Reviewed By: mdko

Differential Revision: D56324033

fbshipit-source-id: a51823c2749db086316e4134ec7dfd2b45d4e99c