Private access to static member through object instance must be from same class as instance
Summary:
Hack currently accepts the following:
```
class C {
  public function testit(D $d): void {
    $d::foo();
  }
}
final class D {
  private static function foo(): void { }
}
```
In an earlier attempt to fix this (D56329255), we simply outlawed access through an instance. Unfortunately it appears to be quite common, and is safe if the class of the instance is the *same* as the enclosing class *and* is `final`. Hence the fix here, which would accept the code in `testit` if it appears inside the final class `D`.

Reviewed By: madgen

Differential Revision: D56348680

fbshipit-source-id: cac77fc7a6df3c882687fe855d0686fe786a0557