Fix broken walkstk for hhvm cores
Summary:
We fix the walkstk command which doesn't show HHVM frames specifically in
coredumps.

Underlying issue is a likely bug in lldb that returns incorrect ordering of the
`HPHP::ActRec` type when looked up from a frame via the `FindTypes` and
`FindFirstType()` APIs. We get a typedef'd type instead of the actual struct
type we want.

This diff fetches the correct type by iterating over all the found `ActRec` types
and picking the correct one and then using that instead.

We also switch from using `unsigned_cast(fp, $TYPE)` to `fp.Cast($TYPE)`
because the typecast SBValue returned by unsigned_cast is unable to actually
dereference and load the HHVM frame into the HPHP::ActRec struct when we need,
which is potentially another LLDB bug.

Reviewed By: mdko

Differential Revision: D56334847

fbshipit-source-id: 7fb1105fada7e2a9c548652132631ddd67f717e6