Fix int64 handling sqlite3 (#9441)
Summary:
The extension was using the sqlite3_*_int API, which would truncate 64-bits
int to 32-bits on some code paths. The fix is to use the 64-bits API variants
constantly throughout the extension.

Added a test that manipulates int64 that are not
representable as int32 to make sure we handle
those correctly.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9441

Reviewed By: paulbiss

Differential Revision: D56295308

Pulled By: ricklavoie

fbshipit-source-id: a8b0ff34dd822fcced4053b47f07b76d0fd30997