Don't crash when doing `hh --lint path/to/dir`
Summary:
As reported by a user in T183238870, calling `hh --lint` with a directory crashes hh.

Ideally, I guess `hh --lint` should just support directories (and walk them or something) but boy that might be a lot of work, and that's too much for an oncall at 4pm on a Friday

This diff just makes sure we tell the user that we don't accept directory names rather than crash on them.

Reviewed By: madgen, viratyosin

Differential Revision: D56355777

fbshipit-source-id: 6dd201fc4247780ca91df3b67d8bbd5eb7172250