Use IntrusiveSharedPtr for HandlerCallback for oneway functions
Summary:
This diff changes `HandlerCallbackBase::Ptr` to use `IntrusiveSharedPtr` which was added in {D56043656}.

This change is useful because it allows shared ownership of the `HandlerCallback` object between user code and Thrift's API.

This approach is being considered after {D55506070} was reverted. It's also a stepping stone on the path to implementing T184318452.

Reviewed By: sethdelliott

Differential Revision: D56043655

fbshipit-source-id: 5ac85f72998203f074b0b675438c005ebf29a768