Generate correct actual type under nullable array access
Summary:
D51114177 modified typing of nullable array get to move inside the `Toption _`. However, the `Hole` that is constructed on error refers to the inner type of the option rather than the original type.

This was manifesting in T180105749 where `mixed` was being turned into `nonull` but we have the same thing where ever there is a type error in the inner type.

Reviewed By: madgen

Differential Revision: D56485872

fbshipit-source-id: 3fd6171e8a6ec635973427b82b2651959d050f9f