enable nesting
Summary:
Allow expression tree expressions inside of splices, as an experimental feature under the `<<file: __EnableUnstableFeatures('expression_tree_nest')>>` attribute.

Prohibit the nested expression tree from having a DSL in front of the first `` ` ``, since it can be taken from the enclosing tree. However, if a splice contains a (hack) lambda, then the body of the lambda is considered a new scope, and top-level expression trees there need the DSL. In contrast, client lambdas in the expression tree do not introduce a new scope wrt splices in their body, and so further nesting does not have the DSL  (see the new test cases for examples).

Nested expression trees do not see (client) variables from enclosing scopes. We plan to improve this in the future.

Reviewed By: viratyosin

Differential Revision: D56326820

fbshipit-source-id: ef34a4bfbc20b4acd8861e6f8cb623ecbda05628