Model MemoGet's possible values correctly in HHBBC
Summary:
When HHBBC models the MemoGet bytecode, it tries to determine the
types that can come out of the memoize cache. In order to do this, it
assumes the memoize impl will be called a certain way and looks up the
function's return type.

Unfortunately this logic assumed that a *non-forwarding* call will
always be used. That is, the variant of FCall that does not forward
the LSB context. However, there are cases where a forwarding variant
is used (if the memoized function comes from a trait). Arguably this
is a bug, but for now HHBBC will match runtime behavior.

As a result of this mismatch, HHBBC could infer a different type from
the memoize cache versus reality. In some cases (for example if the
class is abstract), it could infer that the memo cache will never have
a value in it, and therefore it produces a Bottom!

To fix this, model the memo cache types different. Track the specific
types provided to the memo cache via MemoSet directly, rather than
trying to infer them indirectly. Since the MemoGet should proceed the
MemoSet in RPO order, whenever we update the memo cache information,
we need to reschedule the blocks with the MemoGets in them.

Reviewed By: voorka

Differential Revision: D56534664

fbshipit-source-id: eb7feccbccf2950abda35162503ff37a44f94792