Remove peeking logic for TLS connections
Summary:
For TLS connections, we expect clients to provide an appropriate ALPN. For "rs", "h2", and now "thrift", we'll find a corresponding `TransportRoutingHandler` that we'll use to parse the connection's content. We no longer need to peek at the content. We still have some clients that aren't sending an ALPN value (eg. legacy Java). For those, we can fallback to using Header.

Remove the `PreReceivedDataAsyncTransportWrapper` that was meant to replay the bytes that we peeked at.

In the non-TLS case, we can make the bytes available again to the `AsyncSocketTransport` through `setPreReceivedData`, no need to replay them as part of subsequent read.

Reviewed By: AjanthanAsogamoorthy, praihan

Differential Revision: D44266996

fbshipit-source-id: c9581c65c7b5487f677f0684ebbcf47e121b4b4f