Delete *Exn GetTType impls
Summary:
The point of `GetTType` is for providing type information when serializing collections. `p.write_list_begin(T::TTYPE, len)`

https://www.internalfb.com/code/fbsource/[d49f493bb2852ebef7057ef0c912728b8ec8025e]/fbcode/thrift/lib/rust/src/serialize.rs?lines=291-296%2C298-306

The function-specific error types (`FooExn`, `FooStreamExn`, `FooResponseExn`) are not used in this manner. Only a single one of them is ever serialized in a particular response, and type info isn't needed.

Reviewed By: Imxset21, shayne-fletcher

Differential Revision: D56617794

fbshipit-source-id: 586a2a054697d6fe810bbd46ff15f79d72f726f0