Include new cursor position in snapshot tests
Summary:
D56480226 introduces logic to move the cursor position to inbetween the braces after an "override method" quickfix.

This diff updates the snapshot tests to render that cursor position as `>||<`. For example:

```
...
  public function foo(): void {>||<}
...
```

If the selection would be a nonempty range, the selected characters would appear between `>|` and `|<` like so:

```
some >|selected|< text
```

but currently nothing in `hh` exhibits such behaviour

Reviewed By: mheiber

Differential Revision: D56570133

fbshipit-source-id: dcd0a1c7a65e5cbcc1b791dbc27dfd7112e2b3d5