Don't resolve functions in HHBBC when analyzing constants
Summary:
The point of the analyze constants phase is to fold constants into one
another. So, analyzing functions and their return types, or trying to
inline interp functions is a waste of time. It doesn't really affect
the folded constants, and generates more false dependencies, which
slows down the phase. Avoid all this by not resolving functions or
doing inline interp when we're analyzing constants.

Reviewed By: mdko

Differential Revision: D56619964

fbshipit-source-id: 73a0a6b670dc15e6f4095c9bd924d66df09c0890