add reflection in auto_migration mode
Summary:
This verifies that basic types reflection with thrift-python types works in auto-migrate mode.

Unfortunately, it only works for the subset of types without a container field. We should prioritize generating containers, or keeping them in some limited capacity, because they're going to be a blocker for most modules features fully working

Reviewed By: yoney

Differential Revision: D56601461

fbshipit-source-id: 1e87b2911352580c06e4aae738cd2bec83c4f4d8