Refactor collection of root metadata to enable future resuse in structured logging
Summary: Collecting some metadata from the root for logging is fairly involved. This change does not address that. :) However, we will be adding a new way to log some of this collected data via structured logging and the existing mechanism would not readily support this. To address, we will collect the root metadata into a new struct (`RootMetadata`) and then add that to `PerfSample`, where used. Future diffs will use the same to add to structured logging events.

Reviewed By: MichaelCuevas

Differential Revision: D56578021

fbshipit-source-id: ac664595c8c76c8b9970edff8467dec7c7b0325c