Load object base from the local in LIterNext
Summary:
Since the introduction of IterBase, it is guaranteed that LIterInit/LIterNext
never receive a base that needs to be resolved further. This was previously
affecting objects, as collections could be resolved into the underlying arrays.

So, if LIterNext receives an object via the local, it is guaranteed it is the
same Iterator object as the one stored in the iter slot. So use it instead.

The next diff will stop storing it into the iter slot and clean up related
logic.

JIT implementation will be improved in future diffs to issue calls to rewind(),
next(), valid(), key() and current() directly from JIT. This work is deferred
to better separate cost changes for Iter-LIter merge vs other optimizations.

Reviewed By: ricklavoie

Differential Revision: D56685937

fbshipit-source-id: 9b2a8507043d6b8fb27aba4e288807d0ad04b309