Strip down dependencies of Typing_classes_heap
Summary:
Typing_classes_heap will likely continue to be used by zoncolan, because it contains useful logic to get all sorts of info from folded classes. In the effort to untangle zoncolan from hack, we start stripping down dependencies from Typing_classes_heap. We do that by moving heap-related functions directly to the decl provider (which will not be used anymore by zoncolan). These functions really belong there anyway, and moving them allows us to defunctionalize them. As a result, Typing_classes_heap is not about the heap anymore, it's only about querying information from folded classes. We shall therefore rename it in a subsequent diff.

The only difficulty in doing so is that `Typing_classes_heap.class_t` is abstract. To overcome this, the only thing needed is a wrapper `make_class_t: Typing_class_types.class_t -> class_t`. This is fine because the main purpose of abstracting the type is to prevent looking in the data structure directly.
This wrapper is applied at the boundary of Decl_provider, i.e. right before returning a folded class from `get_class`. As a result, the local cache now contains the non-abstract version of class_t from Typing_class_types.

Differential Revision: D56699704

fbshipit-source-id: 08c34cfe7ce98df1fd3fdf568eebbeefffbf7410