Rename Typing_classes_heap to Folded_class and simplify signature
Summary:
With previous diff D56699704, Typing_classes_heap has nothing to do with the heap anymore, and everything to do with querying information about folded classes. So we simply rename it Folded class.

The Api submodule is now the only top-level definition of the module, so it has no reason to be anymore and we remove it too.

Module Decl_provider.Class is also removed, and replaced with Folded_class.

Most of this diff is mechanical changes. Files to review are:
- folded_class.ml(i)
- decl_provider.ml(i)
- typing_env.ml(i), as an example for other file changes

Differential Revision: D56701920

fbshipit-source-id: a776b12a9a35e02adc367864cd6687515e9340f0