Add HHBBC logic to track locally resolved or context sensitive type-structures
Summary:
A context sensitive type-structure is one that uses "this". Such
type-structures can resolve to different values depending on where the
type-structure is being evaluated. This is important because it
controls how we can re-use any resolved type-structures. If a resolved
type-structure isn't context sensitive, then it can be freely used
regardless of the context. If it is, then it can only be used safely
fromm the type-constant's declaring class.

If a type-structure was resolved in the current worker job, then mark
it with a "locally resolved" bit. This is important to determining who
can use that information across multiple workers.

Also do some cleanup. This is all in preparation for the diff where
this information will be used.

Reviewed By: jano

Differential Revision: D56674357

fbshipit-source-id: b424039643abdcc5c09db628d57c0b31a1ff731c