Improve encoder eviction algorithm and datastructure use
Summary:
The QPACK encoder needs to track what entries in its table are possibly in use by the peer to prevent evicting them too early.  The existing code keeps a reference count for every header in the table, and only evicts when the ref count is 0.  When sending a block referencing tht table, the encoder keeps a list of the absolute table indexes that block is referencing, so it can do the accounting.

This is expensive in both space and time, and unecessary.

We only need to know the minimum value in the table that is referenced, since eviction is in FIFO order.  So we get rid of the (relatively large) refcount table and just keep a minInUseIndex.

The encoder only needs to track the min index for each outstanding block.  The trick is computing the new min efficiently after a header acknowledgement or reset.  Since std::heap functions do not provide a function for removal of an arbitrary element, QPACKEncoder keeps a vector of mins, and the min-min.  Insert (encoding) is O(1).  Removing on ack is O(n), where n is very commonly *at most* 100 and frequently much-much smaller, and the constant factor is vector traversal.  The removal and recomputation of the min is handled in one pass.

Reviewed By: jbeshay

Differential Revision: D56500544

fbshipit-source-id: 476024c43dd3d11d86d60ef76b9773052cad5ece