Don't crash watchman when detected Eden root is not connected (pt. 2)
Summary: Previous attempt, D56043760, fixed a potential crash (when attempting to read `.../.eden/root`), but missed the earlier crash when reading (`.../.eden/socket`). It also missed that the thrown exception (`TerminalWatcherError`) does not have the same serialization support for other clients and would be caught and rethrown as a `std::system_error`. Let'f fix both of those issues by `throwing RootNotConnectedError`'s and catching/rethrowing in the `initWatcher` logic to propogate the error to the caller.

Reviewed By: kmancini

Differential Revision: D56780919

fbshipit-source-id: b6f47d858d57b98b2dbbf632424b7084757324f9