Perform COW optimizations on LIters
Summary:
Perform LIter to LIter optimization similar to the already existing Iter to
LIter optimization. If the underlying local has an equivalent local with a
lower ID (common with LIters over an unnamed local that was copied from a named
local) and all write operations are safe, use it instead.

Differential Revision: D56801410

fbshipit-source-id: b3246be66d766cdaab8a4deba25a2415561d7f54