Break dependency between Decl_enforceability and Decl_provider_internal
Summary:
This is part of the zoncolan detangle.
Zoncolan will likely still use Decl_enforceability but should not use Decl_provider_internal, so we break the dependency.

This is relatively easy since Decl_enforceability is already functorized. We just move the functor application and everything that depends on it to a different module Decl_enforceability_shallow.

Differential Revision: D56760242

fbshipit-source-id: 7ec2d703cb1deb26b158d79b616b5bdcbd48f735