Make "Add method stub" refactoring available in between method definitions (a.o.)
Summary:
This is a handy refactor! People like it. So, I wanna make it available at more locations!

This diff makes the refactor available when you do Cmd+. on whitespace in-between class body elements, such as method definitions, property definitions, etc.

It will insert the selected method in the whitespace where the cursor is positioned.

Reviewed By: enetsee

Differential Revision: D55536820

fbshipit-source-id: eb7f273cd70e26e5ad5ea3a608eea2ee0ec2935a