Support special <<__IgnoreReadonlyError>> attribute for parameter checking
Summary:
For our typed mocking APIs, the various checks on `readonly` on function types (parameters, return, and closed-over variables) get in the way of supplying a lambda that has benign "non-readonly" behaviour.

For this very special case only, we can support an attribute `<<__IgnoreReadonlyError>>` on the lambda parameter that disables the `readonly` checks in subtyping. The implementation relies on having a known function type in hand at the point the parameter is checked. For mocking applications, this will be the case.

Reviewed By: madgen

Differential Revision: D56717395

fbshipit-source-id: 7237c7dbcac2f08975415c716e5ece8233ae8f81