Detect cycles when checking Has_type_member bounds
Summary:
In `Has_type_member.simplify` when the LHS is a generic, we'll use `Common.simplify_generic_l` which will check if any of the upper bounds of the generic satisfy the constraint. If the generic is upper-bounded by another generic which is upper-bounded by the first generic (i.e. `T1 = T2`) then this will result in an infinite recursion.

This diff adds a set of internal types to the subtyping environment in order to detect such cycles and failing. Failing on such a cycle will allow the `try_bounds` in `simplify_generic_l` to move on.

The original subtyping check comes from doing an intersection to avoid redundant bounds when adding bounds to a type parameter generic (`Subtype_simplify.try_intersect_i` in `Type_parameter_env.add_upper_bound` https://fburl.com/code/y4fwdu7a)

 ---

`VisitedGoalsInternal` isn't specific to the `Has_type_member` constraint. It could be used for other constraints that might be susceptible to cycles, but I'll leave that for later diffs if applicable.

Reviewed By: andrewjkennedy

Differential Revision: D56730950

fbshipit-source-id: ab62d04b1bc5fd5fc2e7cde417b8fd52f39f74cd