loosen type-info comparison under libc++
Summary:
With libc++, we may get address-comparison for RTTI objects. If the linker does not fully merge all such symbols across all loaded .so's / dylibs, we get false mismatches. It is better in the case of UniqueInstance to err on the side of false matches than to err on the side of false mismatches.

https://github.com/llvm/llvm-project/blob/llvmorg-18.1.4/libcxx/include/typeinfo#L117-L166

Reviewed By: bcardosolopes

Differential Revision: D56808919

fbshipit-source-id: 453cac1f5f1e48c62394fa272b00d89f10675f54