Coalesce request contexts for related tasks in SerialExecutor
Summary:
`SerialExecutor`s are usually associated with a request, so all tasks enqueued in them generally share the same request context. We can optimize for this case by not resetting the previous request context on each task, and instead calling `setContext()` directly, which internally checks if the new context is the same and skips the `RequestData` callbacks.

Even if the context is not the same, but it shares some `RequestData` (for example if `ShallowCopyRequestContextScopeGuard` was used), `setContext` will avoid the `onUnset()/onSet()` pairs on those `RequestData`.

In practice, if the parent executor correctly handles the request context this diff is a no-op because the worker task will already run under the request context of its task, but I would like to make this explicit as I plan to apply similar optimizations down in the stack.

Reviewed By: iahs

Differential Revision: D56818685

fbshipit-source-id: 89ccd9ad4e0b36300085a9446c465347d8330bd5