use a side map to store string memokey to IC, use int as the memo key
Summary:
This was an idea from jano while discussing IC performance.

The biggest cost of using IC's actually come from the memoization sharding part where we are inserting the entire IC's key in the memo cache. This problem gets worse when we have multiple KeyedByIC's in a PHP stack.

If we can somehow compress the IC key, we can save that cost.

**How does IC memokey work?**

IC memo keys for keyedByICs are generated by appending the current IC to "everything we have so far".

Example1: Current state is Inaccessible, and we are going into KeyedByIC, we use the static key for inaccessible which is %Inaccessible%, then append the current object memokey on (which itself contains the class + the IC's key).

Example2: Suppose we are going from KeyedByIC to another keyedByIC, we create the new IC, then we take everything we have thus far in the hierarchy of that context, append the current object info, and use that as memokey

**Why do we append the keys to a string this way?**
Just to guarantee the unique memo_key for each IC.

**Alternative?**
If we were to use an integer to identify the IC object, we could save a lot of resources computing this large key every time.

**This diff**

1. Extract the logic of computing Ic's memokey out.
2. Create the IC memokey first before creating an IC.
3. Use a side map to hold the Ic_memo_key inside the IC map
4. When the IC is requested, return it from the map if found. If not, create one and store it in the map
5. Change the memokey type for the IC memokey to be int (the address of the IC object
7. Update the tests

**Why this works** The existing memokey as a string will be the same every time when requested from the same class/func hierarchy. So, as long as map each string to the corresponding object, the address of the object can double as the memo key

Reviewed By: jano

Differential Revision: D56385248

fbshipit-source-id: 5d4b1d6c5514b39abb0f58f2cb530c37317ad841