Add extended reasons for unions
Summary:
Adds extended reasons through union subtyping and distinguishes symmetric and asymmetric projections for this case.

A symmetric projection means we made the same projection on both sub- and supertype where as an asymmetric projection means we project only one.

When we have a union in subtype, we update the provenance to record an (asymmetric) projection into an element of the union and leave the reason for the supertype untouched. The reverse happens when we have a union in supertype position.

We follow the existing rules in subtype constraint simplification to handle unions in both sub- and supertype position

Reviewed By: madgen

Differential Revision: D56703724

fbshipit-source-id: 71a72d3f987a7e909c36fbafafd11b918bc9589b