Don't use reasons to suppress errors
Summary:
A special `Reason.t` is used to indicate a type variable was generated to represent an error case. We then inspect the reason during solving to determine if an  error should be raised or not based on whether this is the attached reason.

This use of reasons becomes trickier under extended reasons since we have to traverse the path.

This  diff moves the marker for type variable representing errors inside [tyvar_info] withing the inference environement and modifies helpers used to generate the error related type variables

Reviewed By: andrewjkennedy

Differential Revision: D56700535

fbshipit-source-id: 413c42d2c599acd2a1c3c0dddd2487c52634c34d