Add extended reasons for assignment to locals & list lvalues
Summary:
When assigning to a local, typing currently gives the typed expression the same reason as the rhs. This diff changes that behavior when `--extended-reasons` is set so that we record a flow from the rhs into the lhs. At the moment, the extended reasons are only recorded for assignments to locals and `list(...)` lvalues.

Since lvalue now have a different reason to the rhs, we also have to update the solver to check for types introduce because of type errors anywhere on the path.

Reviewed By: madgen

Differential Revision: D56632238

fbshipit-source-id: 93b1571ebab2fa353f356032a2d4517b513c13ba