Serialize optional `command` argument to `set_selection` command to be executed afterwards
Summary:
In the previous diff we added the functionality to add an optional `command` to the `hack.setSelection` command in the Hack Nuclide
extension that will be executed after changing the selection.

This diff exposes that functionality from OCaml, but currently no such command is ever included in the `hack.setSelection` command.

Reviewed By: enetsee

Differential Revision: D56671499

fbshipit-source-id: 14df0c3906f39ac580ee389c9b856a91d70b507e