Refactor standard default value functions
Summary:
There are 3 functions:
- `createStructTupleWithDefaultValues()`
- `populateStructTupleUnsetFieldsWithDefaultValues()`
- `createStructTypeInfo()`

These functions utilize `getStandardImmutableDefaultValue()`, and they are invoked by the Cython layer for both immutable and mutable types.

This refactoring introduces wrapper functions for all 3 of these functions. The new wrappers add 'Immutable' to the function names and call the original functions, passing `getStandardImmutableDefaultValue` as a parameter.

This refactoring makes it possible to add `getStandardMutableDefaultValue()` and 'Mutable' wrappers next.

Reviewed By: aristidisp

Differential Revision: D56855010

fbshipit-source-id: f5d823ea9b201227c3b9c031865d4b5c2293da12