Allow using dc for default cert in cert manager
Summary:
We previously had logic not to allow the wildcard match for DCs in the cert manager. This was to avoid the following
1. Cert matches wild card match for DC
2. Cert also matches a more specific match for non DC (regular cert).
In the above case we'd match on (1)  if we didn't have this additional MatchType::Direct check

Instead I plan to put the onus on the creator of the CertManager to manage this problem. Note that the caller must pass in the defaultCert = true value to addDelegatedCredential, so this shouldn't affect existing users.

Reviewed By: sotodel

Differential Revision: D55970432

fbshipit-source-id: 9a111b9eeb723b04d335eca29a08f153550716f0