Introduce `getStandardMutableDefaultValueForType()`
Summary:
Introduce the `getStandardMutableDefaultValueForType()` function for mutable types as a counterpart to the `getStandardImmutableDefaultValueForType()` function. The standard default value for a type is defined by the `getStandardImmutableDefaultValueForType()` function and it is specific to immutable types. For mutable types, we are going to change them as the implementation progresses. For example, immutable types utilize Python `tuple` for a Thrift list, however, mutable types will utilize Python `list`.
This change copies the `getStandardImmutableDefaultValueForType()` and changes the standard default value from empty `tuple` to empty `list`.

Reviewed By: aristidisp

Differential Revision: D56888790

fbshipit-source-id: 16bbc071fd882d2a61cd6a2183750b8addbf0e97