Refactor getRedirectDestination With Static Public Helper
Summary:
From message:

getRedirectDestination is currently a private method inside `HTTPRedirectHandler`. Given a location header from a 3xx, it is able to output the url.

This is a useful functionality. Refactor the private function `getRedirectDestination` to a public static function in `ParseURL`.

Move the anon function `isSupportedScheme` to `ParseURL` instead.

Reviewed By: hanidamlaj

Differential Revision: D56788027

fbshipit-source-id: 36318459bcc28dd4a788d083d3746446a7bbc54f