type_predicate - add predicates: int, string, arraykey, float, num, resource, null
Summary:
This diff adds more type_predicate kinds and replaces most creations of `Neg_prim` with an analogous `Neg_predicate`.

Main value of this diff is that it susses out some of the issues that arise when you have more `Type_switch` and `Neg_predicate` types which are handled by the new `Type_switch` logic.

Reviewed By: dlreeves

Differential Revision: D56654249

fbshipit-source-id: 75594367b72c0412e368e61e0ba025f5ca730aa9