Implement trace scheduling for HHBBC analysis phase
Summary:
Change the scheduling algorithm in AnalysisScheduler to use "trace
scheduling". The overall idea of trace scheduling is to include all of
a item's transitive dependencies. If one of those transitive
dependencies changes, we'll reprocess it in *every* bucket it's
present in, and propagate the results within the worker. The goal of
this is to reduce the number of analysis passes needed before reaching
a fixed-point. We trade more computation (by having the same thing be
potentially processed multiple times across workers, and in exchange
analysis gets propagated much quicker. This is, in general, a good
trade-off, since workers are relatively expensive to spin up, while
work within the same unit is cheaper.

One wrinkle to this algorithm is that it only works if the analysis of
an item is consistent in every worker. One way they be inconsistent if
the analysis if one item utilizes the presence of another item that
not all workers have. So, we need to introduce a permission model. On
a given worker, it's not enough to see if some data is present. Now
one must also check if they're "allowed" to use that data. This is
accomplished by storing bit sets of all the workers an item is present
in, and doing subset checks on different sets.

To help in debugging (or for the curious), add a facility to dump a
graph of transitive dependencies for a given class/func/unit. This can
be enabled using the HHBBC_TRACE_FUNCS environment variable and
setting TRACE=hhbbc_dump_trace:1.

This new scheduling actually slows down analyze constants a bit, but
the hope is that it will speed-up the full analysis phase.

Reviewed By: jano

Differential Revision: D56780897

fbshipit-source-id: eccb1670cc9ee37f4fb76b22236926c88b49dfe0