Add new DCls variant to HHBBC
Summary:
Add a new DCls flavor to HHBBC. This new flavor is called "isect and
exact". It represents an intersection of classes (all of which are
<=), and a single exact class. At first this seems non-sensical. Why
intersect an exact class with a list of sub classes, when the end
result of that intersection should always be that exact class (or
Bottom)?

This is useful when the exact class is "missing". That is, we don't
know anything about the class (if it even truely exists). Suppose you
intersect Cls=Missing and Cls<={C1&C2}. There's only really two
possible options for the result. Either Missing is a subclass of C1
and C2, or it is not. In the former case, the result should be
Cls=Missing, and in the later case the result should be Bottom. This
is what the type-system has been doing until now.

However, this actually can break invariants on the types. For example,
for intersection we want to assert that the intersection of two types
is a subtype of the two types. However the above intersection breaks
that! Cls=Missing & Cls<={C1&C2} = Cls=Missing. Cls=Missing <=
Cls=Missing, so no problem there, but Cls=Missing is not a subset of
Cls<={C1&C2}! (Remember, we don't know anything about Missing, so we
can't say it's a subtype).

To make this work, use the new flavor of DCls to have the intersection
result in a type Cls={Missing}&{C1&C2}. By keeping both the
intersection list and the exact class, we can now have the subtype
logic realize the result is a subtype of both the inputs! The exact
class must always be missing, and the intersection list members
shouldn't have complete children. If either of those conditions is
broken, we can resolve the intersection more precisely.

With this implemented, a bunch of special casing logic in the HHBBC
type-system tests can be removed. While I'm at it, make the "nonReg"
parameter to subCls and clsExact be explicit, rather than defaulting
to true. Having a default value leads to bugs because it's easy to
forget to specify false when necessary.

Reviewed By: jano

Differential Revision: D56839412

fbshipit-source-id: 803636f138999add96f382d3c19c5a7cd7ac5130