Remove some apparently unneeded Concat aggregation logic in HHBBC
Summary:
I'm not sure what this is trying to accomplish, but it
doesn't seem correct. Moreover, it can break monotonicity. It can
cause other optimizations not to trigger when this done. If we analyze
the func again, this logic won't kick in (because the Concats have
already been combined), and then the other optimizations will then
fire. However this means we're not truely reaching a fixed-point. Just
remove this logic as I cannot figure out what it's trying to
accomplish and everything still appears to work without it.

Reviewed By: mdko

Differential Revision: D56869780

fbshipit-source-id: bd1ea3f2e4c1275b60f1387df8fbc835bf76843f