Store perfmap in core dumps
Summary:
/tmp/perf-${PID}.map file has useful content for figuring out which translation
did we crash in and for correlating main, cold and frozen sections of that
translation.

For example, if the last operation was a side exit to the interpreter, it was
likely in the cold or frozen sections and the code usually does not have back
references to the main section and it involves a lot of hunting to figure it
out. The perfmap file can be used to find the main section.

Reviewed By: mdko

Differential Revision: D56982015

fbshipit-source-id: 997555a58a5efaf08dd2adf71241e60a9f2f2d86