Introduce concept of "serialized" res::Class
Summary:
Serializing Types via BlobEncoder (and decoding via BlobDecoder)
introduces a few issues. It can make a Type "non-canonical" because
the serialized Type reflects the class information the extern-worker
job knew. This class information may be different from another job
knows. Namely if the current job has better class information, it may
have produced a better Type than what the other job produced. This
usually appears with intersection lists. Another job may have produced
a Type which an intersection list, but another job may have had better
class information and thus was able to produce a Type with a single
class.

To fix this, introduce the concept of a "serialized" res::Class. A
serialized res::Class is basically just the name of the class and you
can't do anything with it. If you want to actually use that class, you
must call the unserialize method on it. At that point, it will
actually look up the class and then be usuable. This is usually not
done yourself, but via unserialize_classes in type-system.cpp. This
converts any classes within the Type to not be serialized and brings
the Type to "canonical" form (as far as known class information
permits). Any Type which comes from from BlobDecoder must have
unserialize_classes called on it before being used. This also
registers any dependencies needed, to ensure the Type always
canonicalizes to that or better in later rounds.

Reviewed By: mdko

Differential Revision: D56967225

fbshipit-source-id: 7cc0033c5836088e09b3c99afb2904e77a74aa46