Fix rds::flush() to use system's page size
Summary:
When madvising away the local RDS, the code was hardcoding 4KB for the page size,
which isn't always correct (this hardcoded constant has been around since 2014
-- D1545777). This could result in inadvertently madvising away RDS locals,
including `RequestInfo::s_requestInfo`, causing them to be zeroed out for
following accesses.

There was also a bug computing one of the sizes to be madvised, which is also fixed.

Reviewed By: ricklavoie

Differential Revision: D56976784

fbshipit-source-id: 36998891a8380072f425807096aa63b686ac18a5