Remove recursion from objprof (part 1 / n)
Summary:
The diff just does a tiny refactoring to put cycle detection into a function.

Mostly wanted to describe the approaches I was thinking for this task and get any opinions from others in the team .

**Background**:
`ext_objprof.cpp` gets the size of an arbitrarily nested object with cycle detection and some pruning (if recursion reaches stack size of `150`). It also sometimes puts sizes of object and array properties into a `histogram`. The task is to make the algorithm iterative.

There's 3 main functions

- `getObjSize` which gets the size of `ObjectData*`. The object could be an array, in which case it calls `sizeOfArray`, or if it is a pair it calls `tvGetSize`. In general, it iterates over property members to sum up the sizes using `tvGetSize`. It may put the size of each property into a histogram.
- `sizeOfArray` works on `ArrayData*` and iterates over the array to sum up the sizes. If the array is a "mixed array", it may put the size of each element into a histogram.
- `tvGetSize` works on any ref counted type but if any of the other 2 functions calls this method, then nothing goes inside the histogram inside this function itself. Calling `tvGetSize` on an object is not necessarily the same as calling `getObjSize` because `tvGetSize` doesn't work on root objects.

**Goals**
- Make sure performance is still consistent.
- Keep most of the code the same for the 3 methods described above with small refactoring. Makes it easier to review and see correctness.

**Approach 1**

Make a new class `ObjProfNode` that has fields such as
- Variant of ObjectData*, ArrayData* and TypedValue, corresponding to each of the 3 function calls.
- `size` and `sized` for this object.
- `parentIdx` to indicate where the parent node is in the stack
- An optional `histogram_key`  for what key to use to update histogram.
- `refCount`, this is relevant in `tvGetSize` where the `sized` property of `sizeOfArray` call is normalized.

We could create a root node `ObjProfNode` and put it into a stack. Then the last node in the stack can be in 2 main states.

1. It expanded the children already, in which case its `size` and `sized` is already updated by the children nodes that we processed. We can update histogram if that was requested for this node. And we can update this node's parent sizes.
2. It didn't expand the children yet, so we should add the children to the stack.

General algorithm in pseudocode
```
stack = Stack([objProfRoot])
while !stack.empty():
    node = stack.back()

    if node.expandedChildren:
        if node.histogramKey.exists:
            updateHistogram(node)
        updateParentSizes(stack[node.parentIdx])
        stack.popBack()
    else:
        # Expand children nodes
        if node.isArray:
            # Rename of sizeOfArray. Instead of
            # it making recursive calls, it would add
            # a child `ObjProfNode` to the stack
            visitArray(node, stack)
        elif node.isObject:
            visitObject(node, stack)
        else:
            visitTypedValue(node, stack)
        node.expandedChildren = true
return pair(objProfRoot.size, objProfRoot.sized)
```

The benefit of this approach is I think it's relatively simple. But the space complexity would be the maximum number of children possible in any call path. This is because we expand the children eagerly instead of having an iterator.

**Approach 2**

Same as approach 1 but we use an iterator to expand children in a more lazy way. When we visit a node, we only expand 1 child that we didn't add before. Then the space complexity would be `O(H * sizeof(ObjProfNode)`) instead. The nice thing is then the `stack` would be like a call stack and we don't need `parentIdx`. Downside is I think iterators don't exist for these objects (?).

Reviewed By: mdko

Differential Revision: D57058586

fbshipit-source-id: 8ad1d6711cb7ecc239b48acce391480e3b17c3ea