Standardize when we trust a Class* 1/n
Summary:
Splitting this up to make debugging easier.

This diff unifies `getClassDependency()`, `classof()` and `subtypeOf()`

It looks like main difference is that `getClassDependency()` did not look at instance bits that are set in repo auth mode.

 ---

There are 3 conditions when we can trust a Class*
(1) persistent
(2) we are jitting a translation in the context of a class that is a subtype of the class
(3) we are jitting a function in the same unit as the cls

Reviewed By: binliu19

Differential Revision: D56801227

fbshipit-source-id: 927aa52db6681a0ef105eb76177f28aa0b394ebc