Use per DataType iterator profiling info
Summary:
If the same foreach loop is executed with different base types, we collect
profiling info per each profiling translation guarded by the base. But once we
try to emit an optimized translation containing the same LIterInit with
different base types, we give up and just use the C++ helpers.

Fix this by maintaining the SpecializedIterator state per DataType.

It is also obvious that it doesn't make sense to profile information about the
base type itself, since it is known, so remove that and just grab the type from
the base SSATmp as needed.

One minor difference is that previously we specialized iterations over dicts as
long as >=99% of them did not have tombstones, but now we require no observed
tombstones during profiling. This does not seem to make a difference, as core
helpers such as Dict\map() were used with tombstones more than 1% of the time.
Later diffs will implement jitted iteration support for tombstones and make
this better than before.

Differential Revision: D56966727

fbshipit-source-id: 27344347b76520c143d17af1bd0e1a535ae9cd15