Specialize IterInit even if there is no IterNext
Summary:
If we can emit a specialized IterInit, emit it unconditionally without the
JmpPlaceholder logic. There's no reason to go to C++ logic if we can jit it.

Record SpecializedIterator the first time we do so and continue using it by
IterNext. This will get improved further in later diffs.

Reviewed By: ricklavoie

Differential Revision: D57036884

fbshipit-source-id: daea0bd4848cbbaafb82ab9a21b02fbf6f2a89ba