Fix test timeout for SingletonThreadLocalTest.OneSingletonPerThread
Summary:
Use a latch to keep threads alive rather than an atomic spinlock. Threads need to be kept alive so that SingletonThreadLocal does not re-use objects across threads.

On local dev machine test is now taking ~1.5 seconds, as opposed to ~5.5 seconds.

The test run has been flaky because of timeouts. The atomic spinlock is my best guess as to why it would time out.

Reviewed By: yfeldblum

Differential Revision: D57073363

fbshipit-source-id: 65d9130e7f8d4209df1e0678e99a1fc8a4816778